[Ads] renamed files
