+++ MaterialsLib/Adsorption/ReactionInert.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "reaction.h"
+#include "Reaction.h"
 
 namespace Ads
 {
